# Add a CUDA kernel for fusing mapping and weighted sum for MoE.