<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

A fusion kernel that improves the overall CUTLASS MOE layer perf for B200.
Fuses this single line: 
`(c2[c_map].view(m, topk, k) * topk_weights.view(m, topk, 1).to(out_dtype)).sum(dim=1)`

## Modifications
Previous perf: https://github.com/sgl-project/sglang/pull/5694
After the change
```
--- Batch Size: 1 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Using default MoE kernel config. Performance might be sub-optimal! Config file not found at /usr/local/lib/python3.12/dist-packages/sglang/srt/layers/moe/fused_moe_triton/configs/E=256,N=256,device_name=NVIDIA_Graphics_Device,dtype=fp8_w8a8,block_shape=[128, 128].json, you can create them with https://github.com/sgl-project/sglang/tree/main/benchmark/kernels/fused_moe_triton
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.069 ms (median) [0.069 - 0.069]
Triton  fused_experts time: 0.128 ms (median) [0.128 - 0.128]

--- Batch Size: 4 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.107 ms (median) [0.107 - 0.107]
Triton  fused_experts time: 0.186 ms (median) [0.186 - 0.186]

--- Batch Size: 16 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.297 ms (median) [0.296 - 0.297]
Triton  fused_experts time: 0.334 ms (median) [0.333 - 0.337]

--- Batch Size: 32 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.465 ms (median) [0.460 - 0.487]
Triton  fused_experts time: 0.495 ms (median) [0.476 - 0.526]

--- Batch Size: 64 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.539 ms (median) [0.536 - 0.624]
Triton  fused_experts time: 0.630 ms (median) [0.603 - 0.654]

--- Batch Size: 128 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.669 ms (median) [0.660 - 0.697]
Triton  fused_experts time: 0.738 ms (median) [0.639 - 0.762]

--- Batch Size: 256 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.867 ms (median) [0.809 - 0.925]
Triton  fused_experts time: 0.762 ms (median) [0.667 - 0.800]

--- Batch Size: 512 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.889 ms (median) [0.810 - 0.924]
Triton  fused_experts time: 0.754 ms (median) [0.701 - 0.807]

--- Batch Size: 1024 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 0.895 ms (median) [0.847 - 0.898]
Triton  fused_experts time: 0.826 ms (median) [0.772 - 0.920]

--- Batch Size: 2048 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 1.086 ms (median) [1.083 - 1.131]
Triton  fused_experts time: 1.245 ms (median) [1.188 - 1.287]

--- Batch Size: 4096 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 1.796 ms (median) [1.783 - 1.903]
Triton  fused_experts time: 1.966 ms (median) [1.818 - 1.989]

--- Batch Size: 8192 ---
Config: E=256, topk=8, H=7168, I_shard=512, dtype=torch.bfloat16, block_shape=[128, 128]
Warming up...
Benchmarking Cutlass fused_experts...
Benchmarking Triton fused_experts...
Cutlass fused_experts time: 3.092 ms (median) [3.007 - 3.200]
Triton  fused_experts time: 3.371 ms (median) [3.190 - 3.426]
```
## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.